# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Faregate gateway
  version: 0.1.0
  summary: A permission and payment gateway for AI agents that need onchain data.
  description: |
    Agents buy onchain data by the query. Humans decide what they are allowed to buy.

    The flow an agent follows:

    1. `POST /requests` with a natural-language ask or a structured query. The
       gateway resolves the agent's passport, evaluates a deterministic policy
       and returns a decision and a price. No data is returned here.
    2. If the decision requires a human, wait until `GET /requests/{id}`
       reports `payment_required`.
    3. `GET /data/{id}`. In live mode this answers `402 Payment Required` with
       x402 payment requirements (network `hedera:testnet`, asset USDC). Retry
       with a `PAYMENT-SIGNATURE` header and the data is released with a
       verified receipt. In simulated mode the receipt is stamped `simulated`.

    Every response that carries data or a receipt also carries its provenance.
    Nothing simulated is presented as a chain fact.
servers:
  - url: http://localhost:8402
...
paths:
...
  /data/{id}:
    get:
      tags: [agent]
      summary: Collect the data for an authorised request, paying the fare.
      description: |
        The only route that returns data. Policy is re-evaluated before a price
        is quoted and again before release. In live mode the first call answers
        `402` with x402 requirements; the retry carries `PAYMENT-SIGNATURE`.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Data released.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
        '402':
          description: Payment required. Body and `PAYMENT-REQUIRED` header carry x402 requirements.
        '403':
          description: Refused at the gate. Not authorised, revoked, over budget, or already fulfilled.
        '502':
          description: Payment was taken but the data provider failed. The request is marked failed.
...
components:
  parameters:
...
    RequestId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Address:
      type: string
      pattern: '^0x[a-fA-F0-9]{40}$'
    ResourceKind:
      type: string
      enum: [wallet.balances, wallet.transfers, wallet.activity, protocol.positions, protocol.markets]
    ResourceQuery:
      type: object
      required: [resource, address, lookbackDays]
      properties:
        resource: { $ref: '#/components/schemas/ResourceKind' }
        address: { $ref: '#/components/schemas/Address' }
        lookbackDays: { type: integer, minimum: 1, maximum: 365 }
        protocol: { type: string, pattern: '^[a-z0-9-]{1,40}$' }
...
    PaymentReceipt:
      type: object
      properties:
        requestId: { type: string }
        txHash: { type: string }
        network: { type: string, example: 'hedera:testnet' }
        from: { type: string }
        to: { type: string }
        amount: { type: string, description: Atomic units of the asset. }
        asset: { type: string, example: '0.0.429274' }
        settled: { type: boolean }
        verifiedAt: { type: string, format: date-time }
        verifiedBy: { type: string, enum: [facilitator, rpc, simulated] }
    DataProvenance:
      type: object
      properties:
        provider: { type: string, enum: [the-graph, simulated] }
        source: { type: string }
        queriedAt: { type: string, format: date-time }
        simulated: { type: boolean }
    RequestResult:
      type: object
      properties:
        data: {}
        provenance: { $ref: '#/components/schemas/DataProvenance' }
        analysis: { type: string, description: Grounding-checked model summary. }
        analysisError: { type: string }
        fulfilledAt: { type: string, format: date-time }
...
    DataResponse:
      type: object
      properties:
        requestId: { type: string }
        query: { $ref: '#/components/schemas/ResourceQuery' }
        result: { $ref: '#/components/schemas/RequestResult' }
        payment: { $ref: '#/components/schemas/PaymentReceipt' }
        budget:
          type: object
          properties:
            chargedUsd: { type: number }
            spentTodayUsd: { type: number }
...